chore: removing getInt and using getLong to support more values

+++ samples/AppAmbit.App.Maui.Native.Android/MainActivity.cs
@@ -325,7 +325,7 @@ public class MainActivity : AppCompatActivity
 
         if (discountLabel != null)
         {
-            int discount = AppAmbit.RemoteConfig.GetInt("discount");
+            long discount = AppAmbit.RemoteConfig.GetLong("discount");
             discountLabel.Text = $"{discount}% OFF";
         }
     }
