+++ samples/AppAmbit.App.Maui.Native.macOS/RemoteConfigViewController.cs
@@ -179,7 +179,7 @@ public class RemoteConfigViewController : UIViewController
 
         if (_discountLabel != null)
         {
-            var discount = RemoteConfig.GetInt("discount");
+            var discount = RemoteConfig.GetLong("discount");
             _discountLabel.Text = $"{discount}% OFF";
         }
     }
